requirements.txt
pydantic
google-api-python-client
google-auth-oauthlib
requests
types-requests
types-pillow
piexif
tqdm
types-tqdm
sqlmodel
Pillow
typer
pillow-heif